@@ -15,7 +15,7 @@ platforms:
 
 environment:
   sdk: ">=3.0.0 <4.0.0"
-  flutter: ">=3.0.0 <3.16.0"
+  flutter: ">=3.16.0"
 
 dependencies:
   cached_network_image: ^3.2.1
@@ -25,8 +25,8 @@ dependencies:
   csslib: ^0.17.3
   dropdown_button2: ^2.1.4
   expand_tap_area: ^1.1.0
-  extended_image: ^8.1.0
-  file_picker: ^5.1.0
+  extended_image: ^8.2.0
+  file_picker: ^6.1.1
   flutter:
     sdk: flutter
   flutter_html: ^3.0.0-beta.2
@@ -35,11 +35,12 @@ dependencies:
   flutter_html_table: ^3.0.0-beta.2
   flutter_html_video: ^3.0.0-beta.2
   flutter_image_compress: ^2.0.4
-  flutter_inappwebview: ^5.7.2+3
+  flutter_inappwebview: ^6.0.0
   flutter_layout_grid: ^2.0.5
   flutter_localizations:
     sdk: flutter
-  flutter_quill: 7.4.16
+  flutter_quill: ^9.2.2
+  flutter_quill_extensions: ^9.2.2
   flutter_slidable: ^3.0.0
   flutter_svg: ^2.0.7
   flutter_swipe_action_cell: ^3.0.2
@@ -69,7 +70,7 @@ dependencies:
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  lottie: ^2.1.0
+  lottie: ^3.0.0
   zds_analysis: ^1.0.0
 
 flutter:
