@@ -4,7 +4,7 @@ import 'dart:convert';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_inappwebview/flutter_inappwebview.dart';
-import 'package:flutter_quill/flutter_quill.dart';
+import 'package:flutter_quill/quill_delta.dart';
 
 import '../../../../zds_flutter.dart';
 
@@ -51,7 +51,6 @@ extension DeltaToHtml on Delta {
       </html>
       ''',
       ),
-      initialOptions: InAppWebViewGroupOptions(crossPlatform: InAppWebViewOptions()),
       onWebViewCreated: (InAppWebViewController controller) {
         // JavaScript handler to retrieve the converted HTML and complete the completer.
         controller.addJavaScriptHandler(
@@ -92,7 +91,7 @@ extension HtmlToDelta on String {
     final Completer<Delta> completer = Completer<Delta>();
 
     // Load the Quill JavaScript library from the specified path.
-    final String quillJs = await rootBundle.loadString('packages/zds_flutter/lib/assets/js/quill.min.js');
+    final String quillJs = await rootBundle.loadString('packages/reflexis_ui/lib/assets/js/quill.min.js');
 
     late HeadlessInAppWebView webView;
 
@@ -119,7 +118,6 @@ extension HtmlToDelta on String {
         </html>
         ''',
       ),
-      initialOptions: InAppWebViewGroupOptions(crossPlatform: InAppWebViewOptions()),
       onWebViewCreated: (InAppWebViewController controller) {
         // JavaScript handler to retrieve the converted Delta and complete the completer.
         controller.addJavaScriptHandler(
