@@ -1,6 +1,11 @@
+// ignore_for_file: implementation_imports, public_member_api_docs
 import 'package:flutter/material.dart';
 import 'package:flutter_quill/flutter_quill.dart';
+import 'package:flutter_quill/src/widgets/toolbar/buttons/alignment/select_alignment_buttons.dart';
+import 'package:flutter_quill/src/widgets/toolbar/buttons/arrow_indicated_list_button.dart';
 import 'package:flutter_quill/translations.dart';
+
+import '../../../../zds_flutter.dart';
 
 /// enum for toolbar position
 ///
@@ -82,6 +87,40 @@ enum QuillToolbarOption {
   link,
 }
 
+/// Options for Toolbar buttons
+enum ToolbarButtons {
+  undo,
+  redo,
+  fontFamily,
+  fontSize,
+  bold,
+  subscript,
+  superscript,
+  italic,
+  small,
+  underline,
+  strikeThrough,
+  inlineCode,
+  color,
+  backgroundColor,
+  clearFormat,
+  centerAlignment,
+  leftAlignment,
+  rightAlignment,
+  justifyAlignment,
+  direction,
+  headerStyle,
+  listNumbers,
+  listBullets,
+  listChecks,
+  codeBlock,
+  quote,
+  indentIncrease,
+  indentDecrease,
+  link,
+  search,
+}
+
 /// The default size of the icon of a button.
 const double kDefaultIconSize = 18;
 
@@ -96,20 +135,8 @@ class ZdsQuillToolbar extends QuillToolbar {
   ///Constructor
 
   const ZdsQuillToolbar({
-    super.children = const <Widget>[],
-    super.axis,
-    super.toolbarSize,
-    super.toolbarSectionSpacing,
-    super.toolbarIconAlignment,
-    super.toolbarIconCrossAlignment,
-    super.multiRowsDisplay,
-    super.color,
-    super.customButtons,
-    super.locale,
-    super.afterButtonPressed,
-    super.sectionDividerColor,
-    super.sectionDividerSpace,
-    super.linkDialogAction,
+    super.configurations,
+    required super.child,
     super.key,
   }) : super();
 
@@ -120,13 +147,15 @@ class ZdsQuillToolbar extends QuillToolbar {
     required Set<QuillToolbarOption> toolbarOptions,
     double? toolbarIconSize,
     String? langCode,
-    Color? color,
+    Color? toolbarColor,
   }) {
-    final ColorScheme theme = Theme.of(context).colorScheme;
+    final ColorScheme colorScheme = Theme.of(context).colorScheme;
+    final Color effectiveToolbarColor = toolbarColor ?? colorScheme.surface;
     return ZdsQuillToolbar.basic(
-      toolbarIconSize: toolbarIconSize ?? 24,
+      context: context,
+      toolbarIconSize: toolbarIconSize ?? kDefaultIconSize,
       controller: controller,
-      color: color,
+      color: effectiveToolbarColor,
       locale: (langCode?.isNotEmpty ?? false) ? Locale(langCode!.substring(0, 2)) : null,
       showUndo: toolbarOptions.contains(QuillToolbarOption.undo),
       showRedo: toolbarOptions.contains(QuillToolbarOption.redo),
@@ -150,20 +179,25 @@ class ZdsQuillToolbar extends QuillToolbar {
       showIndent: toolbarOptions.contains(QuillToolbarOption.indentation),
       showQuote: toolbarOptions.contains(QuillToolbarOption.quotes),
       showLink: toolbarOptions.contains(QuillToolbarOption.link),
+      linkRegExp: RegExp(r'(http|ftp|https)://[\w-]+(\.[\w-]+)+([\w.,@?^=%&amp;:/~+#-]*[\w@?^=%&amp;/~+#-])?'),
       showSubscript: false,
       showSuperscript: false,
       showSearchButton: false,
       multiRowsDisplay: false,
       iconTheme: QuillIconTheme(
-        iconSelectedFillColor: theme.secondary,
-        iconUnselectedFillColor: theme.surface,
-        iconUnselectedColor: theme.onSurface,
+        iconButtonUnselectedData: IconButtonData(
+          color: effectiveToolbarColor.onColor,
+        ),
+        iconButtonSelectedData: IconButtonData(
+          color: colorScheme.secondary,
+        ),
       ),
     );
   }
 
   /// Basic constructor for ZdsQuillToolbar
   factory ZdsQuillToolbar.basic({
+    required BuildContext context,
     required QuillController controller,
     Axis axis = Axis.horizontal,
     double toolbarIconSize = kDefaultIconSize,
@@ -202,7 +236,7 @@ class ZdsQuillToolbar extends QuillToolbar {
     bool showSearchButton = true,
     bool showSubscript = true,
     bool showSuperscript = true,
-    List<QuillCustomButton> customButtons = const <QuillCustomButton>[],
+    List<QuillToolbarCustomButton> customButtons = const <QuillToolbarCustomButton>[],
 
     /// Toolbar items to display for controls of embed blocks
     List<EmbedButtonBuilder>? embedButtons,
@@ -270,400 +304,437 @@ class ZdsQuillToolbar extends QuillToolbar {
       showLink || showSearchButton,
     ];
 
-    // //default font size values
-    // final fontSizes =
-    //     fontSizeValues ?? {'Small'.i18n: 'small', 'Large'.i18n: 'large', 'Huge'.i18n: 'huge', 'Clear'.i18n: '0'};
-    //
-    // //default font family values
-    // final fontFamilies = fontFamilyValues ??
-    //     {
-    //       'Sans Serif': 'sans-serif',
-    //       'Serif': 'serif',
-    //       'Monospace': 'monospace',
-    //       'Ibarra Real Nova': 'ibarra-real-nova',
-    //       'SquarePeg': 'square-peg',
-    //       'Nunito': 'nunito',
-    //       'Pacifico': 'pacifico',
-    //       'Roboto Mono': 'roboto-mono',
-    //       'Clear'.i18n: 'Clear',
-    //     };
-
-    //default button tooltips
-    final Map<ToolbarButtons, String> buttonTooltips = tooltips ??
-        <ToolbarButtons, String>{
-          ToolbarButtons.undo: 'Undo'.i18n,
-          ToolbarButtons.redo: 'Redo'.i18n,
-          ToolbarButtons.fontFamily: 'Font family'.i18n,
-          ToolbarButtons.fontSize: 'Font size'.i18n,
-          ToolbarButtons.bold: 'Bold'.i18n,
-          ToolbarButtons.subscript: 'Subscript'.i18n,
-          ToolbarButtons.superscript: 'Superscript'.i18n,
-          ToolbarButtons.italic: 'Italic'.i18n,
-          ToolbarButtons.small: 'Small'.i18n,
-          ToolbarButtons.underline: 'Underline'.i18n,
-          ToolbarButtons.strikeThrough: 'Strike through'.i18n,
-          ToolbarButtons.inlineCode: 'Inline code'.i18n,
-          ToolbarButtons.color: 'Font color'.i18n,
-          ToolbarButtons.backgroundColor: 'Background color'.i18n,
-          ToolbarButtons.clearFormat: 'Clear format'.i18n,
-          ToolbarButtons.leftAlignment: 'Align left'.i18n,
-          ToolbarButtons.centerAlignment: 'Align center'.i18n,
-          ToolbarButtons.rightAlignment: 'Align right'.i18n,
-          ToolbarButtons.justifyAlignment: 'Justify win width'.i18n,
-          ToolbarButtons.direction: 'Text direction'.i18n,
-          ToolbarButtons.headerStyle: 'Header style'.i18n,
-          ToolbarButtons.listNumbers: 'Numbered list'.i18n,
-          ToolbarButtons.listBullets: 'Bullet list'.i18n,
-          ToolbarButtons.listChecks: 'Checked list'.i18n,
-          ToolbarButtons.codeBlock: 'Code block'.i18n,
-          ToolbarButtons.quote: 'Quote'.i18n,
-          ToolbarButtons.indentIncrease: 'Increase indent'.i18n,
-          ToolbarButtons.indentDecrease: 'Decrease indent'.i18n,
-          ToolbarButtons.link: 'Insert URL'.i18n,
-          ToolbarButtons.search: 'Search'.i18n,
-        };
-
-    return ZdsQuillToolbar(
-      key: key,
+    final componentString = ComponentStrings.of(context);
+    final configurations = QuillSimpleToolbarConfigurations(
+      controller: controller,
       axis: axis,
+      toolbarSize: toolbarIconSize * 3,
       color: color,
-      toolbarSize: toolbarIconSize * 2,
       toolbarSectionSpacing: toolbarSectionSpacing,
       toolbarIconAlignment: toolbarIconAlignment,
       toolbarIconCrossAlignment: toolbarIconCrossAlignment,
       multiRowsDisplay: multiRowsDisplay,
-      customButtons: customButtons,
-      locale: locale,
-      afterButtonPressed: afterButtonPressed,
-      children: <Widget>[
-        if (showUndo)
-          HistoryButton(
-            icon: Icons.undo_outlined,
-            iconSize: toolbarIconSize,
-            tooltip: buttonTooltips[ToolbarButtons.undo],
-            controller: controller,
-            undo: true,
-            iconTheme: iconTheme,
-            afterButtonPressed: afterButtonPressed,
-          ),
-        if (showRedo) ...<Widget>[
-          HistoryButton(
-            icon: Icons.redo_outlined,
-            iconSize: toolbarIconSize,
-            tooltip: buttonTooltips[ToolbarButtons.redo],
-            controller: controller,
-            undo: false,
-            iconTheme: iconTheme,
-            afterButtonPressed: afterButtonPressed,
-          ),
-          if (showDividers) QuillDivider(axis, color: sectionDividerColor, space: sectionDividerSpace),
-        ],
-        if (showHeaderStyle) ...<Widget>[
-          SelectHeaderStyleButton(
-            tooltip: buttonTooltips[ToolbarButtons.headerStyle],
-            controller: controller,
-            axis: axis,
-            iconSize: toolbarIconSize,
-            iconTheme: iconTheme,
-            afterButtonPressed: afterButtonPressed,
-          ),
-          if (showDividers) QuillDivider(axis, color: sectionDividerColor, space: sectionDividerSpace),
-        ],
-        if (showBoldButton)
-          ToggleStyleButton(
-            attribute: Attribute.bold,
-            icon: Icons.format_bold,
-            iconSize: toolbarIconSize,
-            tooltip: buttonTooltips[ToolbarButtons.bold],
-            controller: controller,
-            iconTheme: iconTheme,
-            afterButtonPressed: afterButtonPressed,
-          ),
-        if (showSubscript)
-          ToggleStyleButton(
-            attribute: Attribute.subscript,
-            icon: Icons.subscript,
-            iconSize: toolbarIconSize,
-            tooltip: buttonTooltips[ToolbarButtons.subscript],
-            controller: controller,
-            iconTheme: iconTheme,
-            afterButtonPressed: afterButtonPressed,
-          ),
-        if (showSuperscript)
-          ToggleStyleButton(
-            attribute: Attribute.superscript,
-            icon: Icons.superscript,
-            iconSize: toolbarIconSize,
-            tooltip: buttonTooltips[ToolbarButtons.superscript],
-            controller: controller,
-            iconTheme: iconTheme,
-            afterButtonPressed: afterButtonPressed,
-          ),
-        if (showItalicButton)
-          ToggleStyleButton(
-            attribute: Attribute.italic,
-            icon: Icons.format_italic,
-            iconSize: toolbarIconSize,
-            tooltip: buttonTooltips[ToolbarButtons.italic],
-            controller: controller,
-            iconTheme: iconTheme,
-            afterButtonPressed: afterButtonPressed,
-          ),
-        if (showUnderLineButton)
-          ToggleStyleButton(
-            attribute: Attribute.underline,
-            icon: Icons.format_underline,
-            iconSize: toolbarIconSize,
-            tooltip: buttonTooltips[ToolbarButtons.underline],
-            controller: controller,
-            iconTheme: iconTheme,
-            afterButtonPressed: afterButtonPressed,
-          ),
-        if (showStrikeThrough)
-          ToggleStyleButton(
-            attribute: Attribute.strikeThrough,
-            icon: Icons.format_strikethrough,
-            iconSize: toolbarIconSize,
-            tooltip: buttonTooltips[ToolbarButtons.strikeThrough],
-            controller: controller,
-            iconTheme: iconTheme,
-            afterButtonPressed: afterButtonPressed,
-          ),
-        if (showSmallButton)
-          ToggleStyleButton(
-            attribute: Attribute.small,
-            icon: Icons.format_size,
-            iconSize: toolbarIconSize,
-            tooltip: buttonTooltips[ToolbarButtons.small],
-            controller: controller,
-            iconTheme: iconTheme,
-            afterButtonPressed: afterButtonPressed,
-          ),
-        if (showInlineCode)
-          ToggleStyleButton(
-            attribute: Attribute.inlineCode,
-            icon: Icons.code,
-            iconSize: toolbarIconSize,
-            tooltip: buttonTooltips[ToolbarButtons.inlineCode],
-            controller: controller,
-            iconTheme: iconTheme,
-            afterButtonPressed: afterButtonPressed,
-          ),
-        if (showDividers) QuillDivider(axis, color: sectionDividerColor, space: sectionDividerSpace),
-        if (showColorButton)
-          ColorButton(
-            icon: Icons.color_lens,
-            iconSize: toolbarIconSize,
-            tooltip: buttonTooltips[ToolbarButtons.color],
-            controller: controller,
-            background: false,
-            iconTheme: iconTheme,
-            afterButtonPressed: afterButtonPressed,
-          ),
-        if (showBackgroundColorButton)
-          ColorButton(
-            icon: Icons.format_color_fill,
-            iconSize: toolbarIconSize,
-            tooltip: buttonTooltips[ToolbarButtons.backgroundColor],
-            controller: controller,
-            background: true,
-            iconTheme: iconTheme,
-            afterButtonPressed: afterButtonPressed,
-          ),
-        if (showClearFormat)
-          ClearFormatButton(
-            icon: Icons.format_clear,
-            iconSize: toolbarIconSize,
-            tooltip: buttonTooltips[ToolbarButtons.clearFormat],
-            controller: controller,
-            iconTheme: iconTheme,
-            afterButtonPressed: afterButtonPressed,
-          ),
-        if (embedButtons != null)
-          for (final EmbedButtonBuilder builder in embedButtons)
-            builder(controller, toolbarIconSize, iconTheme, dialogTheme),
-        if (showDividers &&
-            isButtonGroupShown[0] &&
-            (isButtonGroupShown[1] ||
-                isButtonGroupShown[2] ||
-                isButtonGroupShown[3] ||
-                isButtonGroupShown[4] ||
-                isButtonGroupShown[5]))
-          QuillDivider(axis, color: sectionDividerColor, space: sectionDividerSpace),
-        if (showAlignmentButtons)
-          SelectAlignmentButton(
-            controller: controller,
-            tooltips: Map<ToolbarButtons, String>.of(buttonTooltips)
-              ..removeWhere(
-                (ToolbarButtons key, String value) => !<ToolbarButtons>[
-                  ToolbarButtons.leftAlignment,
-                  ToolbarButtons.centerAlignment,
-                  ToolbarButtons.rightAlignment,
-                  ToolbarButtons.justifyAlignment,
-                ].contains(key),
+    );
+    iconTheme ??= QuillIconTheme(
+      iconButtonSelectedData: IconButtonData(
+        iconSize: toolbarIconSize,
+        alignment: Alignment.topCenter,
+      ),
+      iconButtonUnselectedData: IconButtonData(iconSize: toolbarIconSize),
+    );
+
+    return ZdsQuillToolbar(
+      configurations: QuillToolbarConfigurations(
+        sharedConfigurations: QuillSharedConfigurations(
+          locale: locale,
+        ),
+      ),
+      key: key,
+      child: Builder(
+        builder: (context) {
+          //  default button tooltips
+          final buttonTooltips = tooltips ??
+              <ToolbarButtons, String>{
+                ToolbarButtons.undo: componentString.get('STR_UNDO', context.loc.undo),
+                ToolbarButtons.redo: componentString.get('REDO', context.loc.redo),
+                ToolbarButtons.fontFamily: componentString.get('FONT_FAMILY', context.loc.fontFamily),
+                ToolbarButtons.fontSize: componentString.get('FONT_SIZE', context.loc.fontSize),
+                ToolbarButtons.bold: componentString.get('BOLD', context.loc.bold),
+                ToolbarButtons.subscript: componentString.get('SUBSCRIPT', context.loc.subscript),
+                ToolbarButtons.superscript: componentString.get('SUPERSCRIPT', context.loc.superscript),
+                ToolbarButtons.italic: componentString.get('ITALIC', context.loc.italic),
+                ToolbarButtons.small: componentString.get('SMALL', context.loc.small),
+                ToolbarButtons.underline: componentString.get('UNDERLINE', context.loc.underline),
+                ToolbarButtons.strikeThrough: componentString.get('STRIKE_THROUGH', context.loc.strikeThrough),
+                ToolbarButtons.inlineCode: componentString.get('INLINE_CODE', context.loc.inlineCode),
+                ToolbarButtons.color: componentString.get('FONT_COLOR', context.loc.fontColor),
+                ToolbarButtons.backgroundColor: componentString.get('BACKGROUND_COLOR', context.loc.backgroundColor),
+                ToolbarButtons.clearFormat: componentString.get('CLEAR_FORMAT', context.loc.clearFormat),
+                ToolbarButtons.leftAlignment: componentString.get('ALIGN_LEFT', context.loc.alignLeft),
+                ToolbarButtons.centerAlignment: componentString.get('ALIGN_CENTER', context.loc.alignCenter),
+                ToolbarButtons.rightAlignment: componentString.get('ALIGN_RIGHT', context.loc.alignRight),
+                ToolbarButtons.justifyAlignment: componentString.get('JUSTIFY_WIDTH', context.loc.justifyWinWidth),
+                ToolbarButtons.direction: componentString.get('TEXT_DIRECTION', context.loc.textDirection),
+                ToolbarButtons.headerStyle: componentString.get('HEADER_STYLE', context.loc.headerStyle),
+                ToolbarButtons.listNumbers: componentString.get('NUMBERED_LIST', context.loc.numberedList),
+                ToolbarButtons.listBullets: componentString.get('BULLET_LIST', context.loc.bulletList),
+                ToolbarButtons.listChecks: componentString.get('CHECKED_LIST', context.loc.checkedList),
+                ToolbarButtons.codeBlock: componentString.get('CODE_BLOCK', context.loc.codeBlock),
+                ToolbarButtons.quote: componentString.get('EDITOR_QUOTE', context.loc.quote),
+                ToolbarButtons.indentIncrease: componentString.get('INCREASE_INDENT', context.loc.increaseIndent),
+                ToolbarButtons.indentDecrease: componentString.get('DECREASE_INDENT', context.loc.decreaseIndent),
+                ToolbarButtons.link: componentString.get('INSERT_URL', context.loc.link),
+                ToolbarButtons.search: componentString.get('SEARCH', context.loc.search),
+              };
+
+          return Container(
+            decoration: configurations.decoration ?? BoxDecoration(color: configurations.color),
+            constraints: BoxConstraints.tightFor(
+              height: configurations.axis == Axis.horizontal ? configurations.toolbarSize : null,
+              width: configurations.axis == Axis.vertical ? configurations.toolbarSize : null,
+            ),
+            child: Material(
+              color: configurations.color,
+              child: QuillToolbarArrowIndicatedButtonList(
+                axis: Axis.horizontal,
+                buttons: [
+                  if (showUndo)
+                    QuillToolbarHistoryButton(
+                      controller: controller,
+                      isUndo: true,
+                      options: QuillToolbarHistoryButtonOptions(
+                        iconButtonFactor: toolbarSectionSpacing,
+                        iconData: Icons.undo_outlined,
+                        iconSize: toolbarIconSize,
+                        iconTheme: iconTheme,
+                        afterButtonPressed: afterButtonPressed,
+                        tooltip: buttonTooltips[ToolbarButtons.undo],
+                      ),
+                    ),
+                  if (showRedo) ...<Widget>[
+                    QuillToolbarHistoryButton(
+                      controller: controller,
+                      isUndo: false,
+                      options: QuillToolbarHistoryButtonOptions(
+                        iconData: Icons.redo_outlined,
+                        iconSize: toolbarIconSize,
+                        iconTheme: iconTheme,
+                        afterButtonPressed: afterButtonPressed,
+                        tooltip: buttonTooltips[ToolbarButtons.redo],
+                      ),
+                    ),
+                    if (showDividers) QuillToolbarDivider(axis, color: sectionDividerColor, space: sectionDividerSpace),
+                  ],
+                  if (showHeaderStyle) ...<Widget>[
+                    QuillToolbarSelectHeaderStyleButtons(
+                      controller: controller,
+                      options: QuillToolbarSelectHeaderStyleButtonsOptions(
+                        tooltip: buttonTooltips[ToolbarButtons.headerStyle],
+                        axis: axis,
+                        iconSize: toolbarIconSize,
+                        iconTheme: iconTheme,
+                        afterButtonPressed: afterButtonPressed,
+                      ),
+                    ),
+                    if (showDividers) QuillToolbarDivider(axis, color: sectionDividerColor, space: sectionDividerSpace),
+                  ],
+                  if (showBoldButton)
+                    QuillToolbarToggleStyleButton(
+                      attribute: Attribute.bold,
+                      controller: controller,
+                      options: QuillToolbarToggleStyleButtonOptions(
+                        iconData: Icons.format_bold,
+                        iconSize: toolbarIconSize,
+                        tooltip: buttonTooltips[ToolbarButtons.bold],
+                        iconTheme: iconTheme,
+                        afterButtonPressed: afterButtonPressed,
+                      ),
+                    ),
+                  if (showSubscript)
+                    QuillToolbarToggleStyleButton(
+                      attribute: Attribute.subscript,
+                      controller: controller,
+                      options: QuillToolbarToggleStyleButtonOptions(
+                        iconData: Icons.subscript,
+                        iconSize: toolbarIconSize,
+                        tooltip: buttonTooltips[ToolbarButtons.subscript],
+                        iconTheme: iconTheme,
+                        afterButtonPressed: afterButtonPressed,
+                      ),
+                    ),
+                  if (showSuperscript)
+                    QuillToolbarToggleStyleButton(
+                      attribute: Attribute.superscript,
+                      controller: controller,
+                      options: QuillToolbarToggleStyleButtonOptions(
+                        iconData: Icons.superscript,
+                        iconSize: toolbarIconSize,
+                        tooltip: buttonTooltips[ToolbarButtons.superscript],
+                        iconTheme: iconTheme,
+                        afterButtonPressed: afterButtonPressed,
+                      ),
+                    ),
+                  if (showItalicButton)
+                    QuillToolbarToggleStyleButton(
+                      attribute: Attribute.italic,
+                      controller: controller,
+                      options: QuillToolbarToggleStyleButtonOptions(
+                        iconData: Icons.format_italic,
+                        iconSize: toolbarIconSize,
+                        iconTheme: iconTheme,
+                        tooltip: buttonTooltips[ToolbarButtons.italic],
+                        afterButtonPressed: afterButtonPressed,
+                      ),
+                    ),
+                  if (showUnderLineButton)
+                    QuillToolbarToggleStyleButton(
+                      attribute: Attribute.underline,
+                      controller: controller,
+                      options: QuillToolbarToggleStyleButtonOptions(
+                        iconData: Icons.format_underline,
+                        iconSize: toolbarIconSize,
+                        iconTheme: iconTheme,
+                        afterButtonPressed: afterButtonPressed,
+                        tooltip: buttonTooltips[ToolbarButtons.underline],
+                      ),
+                    ),
+                  if (showStrikeThrough)
+                    QuillToolbarToggleStyleButton(
+                      attribute: Attribute.strikeThrough,
+                      controller: controller,
+                      options: QuillToolbarToggleStyleButtonOptions(
+                        iconData: Icons.format_strikethrough,
+                        iconSize: toolbarIconSize,
+                        iconTheme: iconTheme,
+                        afterButtonPressed: afterButtonPressed,
+                        tooltip: buttonTooltips[ToolbarButtons.strikeThrough],
+                      ),
+                    ),
+                  if (showSmallButton)
+                    QuillToolbarToggleStyleButton(
+                      attribute: Attribute.small,
+                      controller: controller,
+                      options: QuillToolbarToggleStyleButtonOptions(
+                        iconData: Icons.format_size,
+                        iconSize: toolbarIconSize,
+                        tooltip: buttonTooltips[ToolbarButtons.small],
+                        iconTheme: iconTheme,
+                        afterButtonPressed: afterButtonPressed,
+                      ),
+                    ),
+                  if (showInlineCode)
+                    QuillToolbarToggleStyleButton(
+                      attribute: Attribute.inlineCode,
+                      controller: controller,
+                      options: QuillToolbarToggleStyleButtonOptions(
+                        iconData: Icons.code,
+                        iconSize: toolbarIconSize,
+                        iconTheme: iconTheme,
+                        afterButtonPressed: afterButtonPressed,
+                        tooltip: buttonTooltips[ToolbarButtons.inlineCode],
+                      ),
+                    ),
+                  if (showDividers) QuillToolbarDivider(axis, color: sectionDividerColor, space: sectionDividerSpace),
+                  if (showColorButton)
+                    QuillToolbarColorButton(
+                      controller: controller,
+                      isBackground: false,
+                      options: QuillToolbarColorButtonOptions(
+                        iconData: Icons.color_lens,
+                        iconSize: toolbarIconSize,
+                        tooltip: buttonTooltips[ToolbarButtons.color],
+                        iconTheme: iconTheme,
+                        afterButtonPressed: afterButtonPressed,
+                      ),
+                    ),
+                  if (showBackgroundColorButton)
+                    QuillToolbarColorButton(
+                      controller: controller,
+                      isBackground: true,
+                      options: QuillToolbarColorButtonOptions(
+                        iconData: Icons.format_color_fill,
+                        iconSize: toolbarIconSize,
+                        iconTheme: iconTheme,
+                        afterButtonPressed: afterButtonPressed,
+                        tooltip: buttonTooltips[ToolbarButtons.backgroundColor],
+                      ),
+                    ),
+                  if (showClearFormat)
+                    QuillToolbarClearFormatButton(
+                      controller: controller,
+                      options: QuillToolbarBaseButtonOptions(
+                        iconData: Icons.format_clear,
+                        iconSize: toolbarIconSize,
+                        iconTheme: iconTheme,
+                        afterButtonPressed: afterButtonPressed,
+                        tooltip: buttonTooltips[ToolbarButtons.clearFormat],
+                      ),
+                    ),
+                  if (embedButtons != null)
+                    for (final EmbedButtonBuilder builder in embedButtons)
+                      builder(controller, toolbarIconSize, iconTheme, dialogTheme),
+                  if (showDividers &&
+                      isButtonGroupShown[0] &&
+                      (isButtonGroupShown[1] ||
+                          isButtonGroupShown[2] ||
+                          isButtonGroupShown[3] ||
+                          isButtonGroupShown[4] ||
+                          isButtonGroupShown[5]))
+                    QuillToolbarDivider(axis, color: sectionDividerColor, space: sectionDividerSpace),
+                  if (showAlignmentButtons)
+                    QuillToolbarSelectAlignmentButtons(
+                      controller: controller,
+                      options: QuillToolbarSelectAlignmentButtonOptions(
+                        tooltips: QuillSelectAlignmentValues(
+                          leftAlignment: buttonTooltips[ToolbarButtons.leftAlignment] ?? context.loc.alignLeft,
+                          centerAlignment: buttonTooltips[ToolbarButtons.centerAlignment] ?? context.loc.alignCenter,
+                          rightAlignment: buttonTooltips[ToolbarButtons.rightAlignment] ?? context.loc.alignRight,
+                          justifyAlignment:
+                              buttonTooltips[ToolbarButtons.justifyAlignment] ?? context.loc.justifyWinWidth,
+                        ),
+                        iconSize: toolbarIconSize,
+                        iconTheme: iconTheme,
+                        showLeftAlignment: showLeftAlignment,
+                        showCenterAlignment: showCenterAlignment,
+                        showRightAlignment: showRightAlignment,
+                        showJustifyAlignment: showJustifyAlignment,
+                        afterButtonPressed: afterButtonPressed,
+                      ),
+                    ),
+                  if (showDirection)
+                    QuillToolbarToggleStyleButton(
+                      attribute: Attribute.rtl,
+                      controller: controller,
+                      options: QuillToolbarToggleStyleButtonOptions(
+                        iconData: Icons.format_textdirection_r_to_l,
+                        iconSize: toolbarIconSize,
+                        iconTheme: iconTheme,
+                        afterButtonPressed: afterButtonPressed,
+                        tooltip: buttonTooltips[ToolbarButtons.direction],
+                      ),
+                    ),
+                  if (showDividers &&
+                      isButtonGroupShown[1] &&
+                      (isButtonGroupShown[2] ||
+                          isButtonGroupShown[3] ||
+                          isButtonGroupShown[4] ||
+                          isButtonGroupShown[5]))
+                    QuillToolbarDivider(axis, color: sectionDividerColor, space: sectionDividerSpace),
+                  if (showDividers &&
+                      showHeaderStyle &&
+                      isButtonGroupShown[2] &&
+                      (isButtonGroupShown[3] || isButtonGroupShown[4] || isButtonGroupShown[5]))
+                    QuillToolbarDivider(axis, color: sectionDividerColor, space: sectionDividerSpace),
+                  if (showListNumbers)
+                    QuillToolbarToggleStyleButton(
+                      attribute: Attribute.ol,
+                      controller: controller,
+                      options: QuillToolbarToggleStyleButtonOptions(
+                        iconData: Icons.format_list_numbered,
+                        iconSize: toolbarIconSize,
+                        iconTheme: iconTheme,
+                        afterButtonPressed: afterButtonPressed,
+                        tooltip: buttonTooltips[ToolbarButtons.listNumbers],
+                      ),
+                    ),
+                  if (showListBullets)
+                    QuillToolbarToggleStyleButton(
+                      attribute: Attribute.ul,
+                      controller: controller,
+                      options: QuillToolbarToggleStyleButtonOptions(
+                        iconData: Icons.format_list_bulleted,
+                        iconSize: toolbarIconSize,
+                        iconTheme: iconTheme,
+                        afterButtonPressed: afterButtonPressed,
+                        tooltip: buttonTooltips[ToolbarButtons.listBullets],
+                      ),
+                    ),
+                  if (showListCheck)
+                    QuillToolbarToggleCheckListButton(
+                      controller: controller,
+                      options: QuillToolbarToggleCheckListButtonOptions(
+                        iconData: Icons.check_box,
+                        iconSize: toolbarIconSize,
+                        iconTheme: iconTheme,
+                        afterButtonPressed: afterButtonPressed,
+                        tooltip: buttonTooltips[ToolbarButtons.listChecks],
+                      ),
+                    ),
+                  if (showCodeBlock)
+                    QuillToolbarToggleStyleButton(
+                      attribute: Attribute.codeBlock,
+                      controller: controller,
+                      options: QuillToolbarToggleStyleButtonOptions(
+                        iconData: Icons.code_off,
+                        iconSize: toolbarIconSize,
+                        iconTheme: iconTheme,
+                        afterButtonPressed: afterButtonPressed,
+                        tooltip: buttonTooltips[ToolbarButtons.codeBlock],
+                      ),
+                    ),
+                  if (showDividers && isButtonGroupShown[3] && (isButtonGroupShown[4] || isButtonGroupShown[5]))
+                    QuillToolbarDivider(axis, color: sectionDividerColor, space: sectionDividerSpace),
+                  if (showQuote)
+                    QuillToolbarToggleStyleButton(
+                      attribute: Attribute.blockQuote,
+                      controller: controller,
+                      options: QuillToolbarToggleStyleButtonOptions(
+                        iconData: Icons.format_quote,
+                        iconSize: toolbarIconSize,
+                        iconTheme: iconTheme,
+                        afterButtonPressed: afterButtonPressed,
+                        tooltip: buttonTooltips[ToolbarButtons.quote],
+                      ),
+                    ),
+                  if (showIndent)
+                    QuillToolbarIndentButton(
+                      controller: controller,
+                      isIncrease: true,
+                      options: QuillToolbarIndentButtonOptions(
+                        iconData: Icons.format_indent_increase,
+                        iconSize: toolbarIconSize,
+                        iconTheme: iconTheme,
+                        afterButtonPressed: afterButtonPressed,
+                        tooltip: buttonTooltips[ToolbarButtons.indentIncrease],
+                      ),
+                    ),
+                  if (showIndent)
+                    QuillToolbarIndentButton(
+                      controller: controller,
+                      isIncrease: false,
+                      options: QuillToolbarIndentButtonOptions(
+                        iconData: Icons.format_indent_decrease,
+                        iconSize: toolbarIconSize,
+                        iconTheme: iconTheme,
+                        afterButtonPressed: afterButtonPressed,
+                        tooltip: buttonTooltips[ToolbarButtons.indentDecrease],
+                      ),
+                    ),
+                  if (showDividers && isButtonGroupShown[4] && isButtonGroupShown[5])
+                    QuillToolbarDivider(axis, color: sectionDividerColor, space: sectionDividerSpace),
+                  if (showLink)
+                    QuillToolbarLinkStyleButton(
+                      controller: controller,
+                      options: QuillToolbarLinkStyleButtonOptions(
+                        iconSize: toolbarIconSize,
+                        iconTheme: iconTheme,
+                        dialogTheme: dialogTheme,
+                        afterButtonPressed: afterButtonPressed,
+                        linkRegExp: linkRegExp,
+                        linkDialogAction: linkDialogAction,
+                        tooltip: buttonTooltips[ToolbarButtons.link],
+                      ),
+                    ),
+                  if (showSearchButton)
+                    QuillToolbarSearchButton(
+                      controller: controller,
+                      options: QuillToolbarSearchButtonOptions(
+                        iconData: Icons.search,
+                        iconSize: toolbarIconSize,
+                        iconTheme: iconTheme,
+                        dialogTheme: dialogTheme,
+                        afterButtonPressed: afterButtonPressed,
+                        tooltip: buttonTooltips[ToolbarButtons.search],
+                      ),
+                    ),
+                  if (customButtons.isNotEmpty)
+                    if (showDividers) QuillToolbarDivider(axis, color: sectionDividerColor, space: sectionDividerSpace),
+                  for (final customButton in customButtons)
+                    QuillToolbarCustomButton(
+                      options: customButton.options,
+                      controller: controller,
+                    ),
+                  for (final customButton in configurations.customButtons)
+                    QuillToolbarCustomButton(
+                      options: customButton,
+                      controller: controller,
+                    ),
+                ],
               ),
-            iconSize: toolbarIconSize,
-            iconTheme: iconTheme,
-            showLeftAlignment: showLeftAlignment,
-            showCenterAlignment: showCenterAlignment,
-            showRightAlignment: showRightAlignment,
-            showJustifyAlignment: showJustifyAlignment,
-            afterButtonPressed: afterButtonPressed,
-          ),
-        if (showDirection)
-          ToggleStyleButton(
-            attribute: Attribute.rtl,
-            tooltip: buttonTooltips[ToolbarButtons.direction],
-            controller: controller,
-            icon: Icons.format_textdirection_r_to_l,
-            iconSize: toolbarIconSize,
-            iconTheme: iconTheme,
-            afterButtonPressed: afterButtonPressed,
-          ),
-        if (showDividers &&
-            isButtonGroupShown[1] &&
-            (isButtonGroupShown[2] || isButtonGroupShown[3] || isButtonGroupShown[4] || isButtonGroupShown[5]))
-          QuillDivider(axis, color: sectionDividerColor, space: sectionDividerSpace),
-        if (showDividers &&
-            showHeaderStyle &&
-            isButtonGroupShown[2] &&
-            (isButtonGroupShown[3] || isButtonGroupShown[4] || isButtonGroupShown[5]))
-          QuillDivider(axis, color: sectionDividerColor, space: sectionDividerSpace),
-        if (showListNumbers)
-          ToggleStyleButton(
-            attribute: Attribute.ol,
-            tooltip: buttonTooltips[ToolbarButtons.listNumbers],
-            controller: controller,
-            icon: Icons.format_list_numbered,
-            iconSize: toolbarIconSize,
-            iconTheme: iconTheme,
-            afterButtonPressed: afterButtonPressed,
-          ),
-        if (showListBullets)
-          ToggleStyleButton(
-            attribute: Attribute.ul,
-            tooltip: buttonTooltips[ToolbarButtons.listBullets],
-            controller: controller,
-            icon: Icons.format_list_bulleted,
-            iconSize: toolbarIconSize,
-            iconTheme: iconTheme,
-            afterButtonPressed: afterButtonPressed,
-          ),
-        if (showListCheck)
-          ToggleCheckListButton(
-            attribute: Attribute.unchecked,
-            tooltip: buttonTooltips[ToolbarButtons.listChecks],
-            controller: controller,
-            icon: Icons.check_box,
-            iconSize: toolbarIconSize,
-            iconTheme: iconTheme,
-            afterButtonPressed: afterButtonPressed,
-          ),
-        if (showCodeBlock)
-          ToggleStyleButton(
-            attribute: Attribute.codeBlock,
-            tooltip: buttonTooltips[ToolbarButtons.codeBlock],
-            controller: controller,
-            icon: Icons.code_off,
-            iconSize: toolbarIconSize,
-            iconTheme: iconTheme,
-            afterButtonPressed: afterButtonPressed,
-          ),
-        if (showDividers && isButtonGroupShown[3] && (isButtonGroupShown[4] || isButtonGroupShown[5]))
-          QuillDivider(axis, color: sectionDividerColor, space: sectionDividerSpace),
-        if (showQuote)
-          ToggleStyleButton(
-            attribute: Attribute.blockQuote,
-            tooltip: buttonTooltips[ToolbarButtons.quote],
-            controller: controller,
-            icon: Icons.format_quote,
-            iconSize: toolbarIconSize,
-            iconTheme: iconTheme,
-            afterButtonPressed: afterButtonPressed,
-          ),
-        if (showIndent)
-          IndentButton(
-            icon: Icons.format_indent_increase,
-            iconSize: toolbarIconSize,
-            tooltip: buttonTooltips[ToolbarButtons.indentIncrease],
-            controller: controller,
-            isIncrease: true,
-            iconTheme: iconTheme,
-            afterButtonPressed: afterButtonPressed,
-          ),
-        if (showIndent)
-          IndentButton(
-            icon: Icons.format_indent_decrease,
-            iconSize: toolbarIconSize,
-            tooltip: buttonTooltips[ToolbarButtons.indentDecrease],
-            controller: controller,
-            isIncrease: false,
-            iconTheme: iconTheme,
-            afterButtonPressed: afterButtonPressed,
-          ),
-        if (showDividers && isButtonGroupShown[4] && isButtonGroupShown[5])
-          QuillDivider(axis, color: sectionDividerColor, space: sectionDividerSpace),
-        if (showLink)
-          LinkStyleButton(
-            tooltip: buttonTooltips[ToolbarButtons.link],
-            controller: controller,
-            iconSize: toolbarIconSize,
-            iconTheme: iconTheme,
-            dialogTheme: dialogTheme,
-            afterButtonPressed: afterButtonPressed,
-            linkRegExp: linkRegExp,
-            linkDialogAction: linkDialogAction,
-          ),
-        if (showSearchButton)
-          SearchButton(
-            icon: Icons.search,
-            iconSize: toolbarIconSize,
-            tooltip: buttonTooltips[ToolbarButtons.search],
-            controller: controller,
-            iconTheme: iconTheme,
-            dialogTheme: dialogTheme,
-            afterButtonPressed: afterButtonPressed,
-          ),
-        // if (showFontFamily)
-        //   QuillFontFamilyButton(
-        //     iconTheme: iconTheme,
-        //     iconSize: toolbarIconSize,
-        //     tooltip: buttonTooltips[ToolbarButtons.fontFamily],
-        //     attribute: Attribute.font,
-        //     controller: controller,
-        //     rawItemsMap: fontFamilies,
-        //     afterButtonPressed: afterButtonPressed,
-        //   ),
-        // if (showFontSize)
-        //   QuillFontSizeButton(
-        //     iconTheme: iconTheme,
-        //     iconSize: toolbarIconSize,
-        //     tooltip: buttonTooltips[ToolbarButtons.fontSize],
-        //     attribute: Attribute.size,
-        //     controller: controller,
-        //     rawItemsMap: fontSizes,
-        //     afterButtonPressed: afterButtonPressed,
-        //   ),
-        if (customButtons.isNotEmpty)
-          if (showDividers) QuillDivider(axis, color: sectionDividerColor, space: sectionDividerSpace),
-        for (final QuillCustomButton customButton in customButtons)
-          if (customButton.child != null) ...<Widget>[
-            InkWell(
-              onTap: customButton.onTap,
-              child: customButton.child,
             ),
-          ] else ...<Widget>[
-            CustomButton(
-              onPressed: customButton.onTap,
-              icon: customButton.icon,
-              iconColor: customButton.iconColor,
-              iconSize: toolbarIconSize,
-              iconTheme: iconTheme,
-              afterButtonPressed: afterButtonPressed,
-              tooltip: customButton.tooltip,
-            ),
-          ],
-      ],
+          );
+        },
+      ),
     );
   }
 }
